Crate fadroma

Fadroma

Macros

contract
Define a smart contract

define_init_message
Instantiation interface.

define_q_messages
Query interface.

define_state_singleton
Define the state singleton.

define_tx_messages
Transaction interface.

implement_init
Instatiation. Either defines or imports an InitMsg, and hooks up your init logic to it. Function body must return the initial value of State.

implement_queries
Query implementations.

implement_transactions
Transaction implementations

message
Define an enum that implements the necessary traits (de/serialization, schema generation, cloning, debug printing, equality comparison)

messages
Define an enum with variants that implement the necessary traits

prelude
Import commonly used definitions that need to be available everywhere in the contract