Module fadroma::admin

Transaction authentication by pre-configured admin address.

Modules

simple

Admin functionality where a new admin can be set by sending a single message by the current admin. Use this when the admin can be another contract.

two_step

Admin functionality where a new admin is set using a two-step process. First the current admin initiates the change by nominating a new admin. Then the nominated address must accept the admin role. Use this when the admin is always a wallet address and not a contract.

Structs

AdminNs

Constants

STORE

Functions

assert

Asserts that the message sender is the admin. Otherwise returns an Err.

init

Initializes the admin module. Sets the messages sender as the admin if address is None. You must call this in your instantiate message.

Attribute Macros

require_admin