Trait fadroma::bin_serde::FadromaDeserialize

pub trait FadromaDeserialize: Sized {
    fn from_bytes<'a>(de: &mut Deserializer<'a>) -> Result<Self>;
}

A type that knows how to create an instance of itself given a stream of raw bytes. Can be derived.

Required Methods

fn from_bytes<'a>(de: &mut Deserializer<'a>) -> Result<Self>

Deserialize into a new instance by reading bytes from the provided Deserializer.