Trait fadroma::composability::MutableStorageWrapper

pub trait MutableStorageWrapper<S: Storage> {
    fn storage_mut(&mut self) -> &mut S;

    fn set<Value: Serialize>(&mut self, key: &[u8], value: &Value) -> UsuallyOk { ... }
    fn set_ns<Value: Serialize>(
        &mut self,
        ns: &[u8],
        key: &[u8],
        value: &Value
    ) -> UsuallyOk { ... }
    fn set_multi_ns<Value: Serialize>(
        &mut self,
        ns: &[&[u8]],
        key: &[u8],
        value: &Value
    ) -> UsuallyOk { ... }
    fn remove(&mut self, key: &[u8]) { ... }
    fn remove_ns(&mut self, ns: &[u8], key: &[u8]) { ... }
    fn remove_multi_ns(&mut self, ns: &[&[u8]], key: &[u8]) { ... }
}

Required Methods

fn storage_mut(&mut self) -> &mut S

Provided Methods

fn set<Value: Serialize>(&mut self, key: &[u8], value: &Value) -> UsuallyOk

fn set_ns<Value: Serialize>(
&mut self,
ns: &[u8],
key: &[u8],
value: &Value
) -> UsuallyOk

Uses a single slice as namespace and it combines it with the key

fn set_multi_ns<Value: Serialize>(
&mut self,
ns: &[&[u8]],
key: &[u8],
value: &Value
) -> UsuallyOk

Uses an array of slices which are then combined according to the following spec: https://github.com/webmaster128/key-namespacing#nesting

fn remove(&mut self, key: &[u8])

fn remove_ns(&mut self, ns: &[u8], key: &[u8])

fn remove_multi_ns(&mut self, ns: &[&[u8]], key: &[u8])