Trait fadroma::composable::Composable

pub trait Composable<S, A, Q>: BaseComposable<S, A, Q> {
    fn set<Value>(&mut self, key: &[u8], value: Value) -> Result<(), StdError>
    where
        Value: Serialize;

    fn set_ns<Value>(
        &mut self, 
        ns: &[u8], 
        key: &[u8], 
        value: Value
    ) -> Result<(), StdError>
    where
        Value: Serialize;

    fn set_multi_ns<Value>(
        &mut self, 
        ns: &[&[u8]], 
        key: &[u8], 
        value: Value
    ) -> Result<(), StdError>
    where
        Value: Serialize;

    fn get<Value>(&self, key: &[u8]) -> Result<Option<Value>, StdError>
    where
        Value: DeserializeOwned;

    fn get_ns<Value>(
        &self, 
        ns: &[u8], 
        key: &[u8]
    ) -> Result<Option<Value>, StdError>
    where
        Value: DeserializeOwned;

    fn get_multi_ns<Value>(
        &self, 
        ns: &[&[u8]], 
        key: &[u8]
    ) -> Result<Option<Value>, StdError>
    where
        Value: DeserializeOwned;

    fn remove(&mut self, key: &[u8]) -> Result<(), StdError>;

    fn remove_ns(&mut self, ns: &[u8], key: &[u8]) -> Result<(), StdError>;

    fn remove_multi_ns(
        &mut self, 
        ns: &[&[u8]], 
        key: &[u8]
    ) -> Result<(), StdError>;

    fn humanize<Value>(
        &self, 
        value: Value
    ) -> Result<<Value as Humanize>::Output, StdError>
    where
        Value: Humanize;

    fn canonize<Value>(
        &self, 
        value: Value
    ) -> Result<<Value as Canonize>::Output, StdError>
    where
        Value: Canonize;
}

Required methods

fn set<Value>(&mut self, key: &[u8], value: Value) -> Result<(), StdError> where
Value: Serialize,

fn set_ns<Value>(
&mut self,
ns: &[u8],
key: &[u8],
value: Value
) -> Result<(), StdError> where
Value: Serialize,

Uses a single slice as namespace and it combines it with the key

fn set_multi_ns<Value>(
&mut self,
ns: &[&[u8]],
key: &[u8],
value: Value
) -> Result<(), StdError> where
Value: Serialize,

Uses an array of slices which are then combined according to the following spec: https://github.com/webmaster128/key-namespacing#nesting

fn get<Value>(&self, key: &[u8]) -> Result<Option<Value>, StdError> where
Value: DeserializeOwned,

fn get_ns<Value>(
&self,
ns: &[u8],
key: &[u8]
) -> Result<Option<Value>, StdError> where
Value: DeserializeOwned,

fn get_multi_ns<Value>(
&self,
ns: &[&[u8]],
key: &[u8]
) -> Result<Option<Value>, StdError> where
Value: DeserializeOwned,

fn remove(&mut self, key: &[u8]) -> Result<(), StdError>

fn remove_ns(&mut self, ns: &[u8], key: &[u8]) -> Result<(), StdError>

fn remove_multi_ns(&mut self, ns: &[&[u8]], key: &[u8]) -> Result<(), StdError>

fn humanize<Value>(
&self,
value: Value
) -> Result<<Value as Humanize>::Output, StdError> where
Value: Humanize,

fn canonize<Value>(
&self,
value: Value
) -> Result<<Value as Canonize>::Output, StdError> where
Value: Canonize,