Module fadroma::ensemble

Test multiple contract interactions using unit tests. Feature flag: ensemble

Re-exports

pub use anyhow;

Structs

BankResponse

Block

ContractEnsemble

This the main type in the system that takes care of registering and executing contracts, keeping the blockchain simulation state and allowing the manipulation of particular parameters such as account funds, blocks or contract state in order to efficiently simulate testing scenarios.

DistributionResponse

EnsembleQuerier

ExecuteResponse

InstantiateResponse

Iter

Iterator that iterates over all responses returned by the various modules in order of execution.

MockEnv

ReplyResponse

StakingResponse

Enums

AttributeError

DistributionOp

EnsembleError

RegistryError

ResponseVariants

StakingOp

Traits

ContractHarness

The trait that allows the ensemble to execute your contract. Must be implemented for each contract that will participate in the shared execution. Usually implemented by calling the respective contract function for each method of the trait by passing down the parameters of the method and calling cosmwasm_std::from_binary() on the msg parameter. It can also be used to implement a mock contract directly.

Type Definitions

AnyResult

EnsembleResult