Enum fadroma::prelude::cosmwasm_std::StdError

#[non_exhaustive]
pub enum StdError {
    GenericErr {
        msg: String,
        backtrace: Option<Backtrace>,
    },
    InvalidBase64 {
        msg: String,
        backtrace: Option<Backtrace>,
    },
    InvalidUtf8 {
        msg: String,
        backtrace: Option<Backtrace>,
    },
    NotFound {
        kind: String,
        backtrace: Option<Backtrace>,
    },
    ParseErr {
        target: String,
        msg: String,
        backtrace: Option<Backtrace>,
    },
    SerializeErr {
        source: String,
        msg: String,
        backtrace: Option<Backtrace>,
    },
    Unauthorized {
        backtrace: Option<Backtrace>,
    },
    Underflow {
        minuend: String,
        subtrahend: String,
        backtrace: Option<Backtrace>,
    },
}

Structured error type for init, handle and query.

This can be serialized and passed over the Wasm/VM boundary, which allows us to use structured error types in e.g. integration tests. In that process backtraces are stripped off.

The prefix “Std” means “the standard error within the standard library”. This is not the only result/error type in cosmwasm-std.

When new cases are added, they should describe the problem rather than what was attempted (e.g. InvalidBase64 is preferred over Base64DecodingErr). In the long run this allows us to get rid of the duplication in “StdError::FooErr”.

Checklist for adding a new error:

Add enum case

Add to PartialEq implementation

Add serialize/deserialize test

Add creator function in std_error_helpers.rs

Regenerate schemas

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

GenericErr

Fields

msg: String

backtrace: Option<Backtrace>

Whenever there is no specific error type available

InvalidBase64

Fields

msg: String

backtrace: Option<Backtrace>

InvalidUtf8

Fields

msg: String

backtrace: Option<Backtrace>

Whenever UTF-8 bytes cannot be decoded into a unicode string, e.g. in String::from_utf8 or str::from_utf8.

NotFound

Fields

kind: String

backtrace: Option<Backtrace>

ParseErr

Fields

target: String

the target type that was attempted

msg: String

backtrace: Option<Backtrace>

SerializeErr

Fields

source: String

the source type that was attempted

msg: String

backtrace: Option<Backtrace>

Unauthorized

Fields

backtrace: Option<Backtrace>

Underflow

Fields

minuend: String

subtrahend: String

backtrace: Option<Backtrace>

Implementations

impl StdError

pub fn generic_err<S>(msg: S) -> StdError where
S: Into<String>,

pub fn invalid_base64<S>(msg: S) -> StdError where
S: ToString,

pub fn invalid_utf8<S>(msg: S) -> StdError where
S: ToString,

pub fn not_found<S>(kind: S) -> StdError where
S: Into<String>,

pub fn parse_err<T, M>(target: T, msg: M) -> StdError where
T: Into<String>,
M: ToString,

pub fn serialize_err<S, M>(source: S, msg: M) -> StdError where
S: Into<String>,
M: ToString,

pub fn underflow<U>(minuend: U, subtrahend: U) -> StdError where
U: ToString,

pub fn unauthorized() -> StdError