Module fadroma::prelude::cosmwasm_std::testing

Structs

BankQuerier

MockApi

MockQuerier

MockQuerier holds an immutable table of bank balances TODO: also allow querying contracts

StakingQuerier

Constants

MOCK_CONTRACT_ADDR

Functions

mock_dependencies

All external requirements that can be injected for unit tests. It sets the given balance for the contract itself, nothing else

mock_dependencies_with_balances

Initializes the querier along with the mock_dependencies. Sets all balances provided (yoy must explicitly set contract balance if desired)

mock_env

Just set sender and sent funds for the message. The rest uses defaults. The sender will be canonicalized internally to allow developers pasing in human readable senders. This is intended for use in test code only.

Type Definitions

MockQuerierCustomHandlerResult

The same type as cosmwasm-std’s QuerierResult, but easier to reuse in cosmwasm-vm. It might diverge from QuerierResult at some point.

MockStorage