Trait fadroma::prelude::cosmwasm_std::Storage

pub trait Storage: ReadonlyStorage {
    fn set(&mut self, key: &[u8], value: &[u8]);
    fn remove(&mut self, key: &[u8]);
}

Required Methods

fn set(&mut self, key: &[u8], value: &[u8])

fn remove(&mut self, key: &[u8])

Removes a database entry at key.

The current interface does not allow to differentiate between a key that existed before and one that didn’t exist. See https://github.com/CosmWasm/cosmwasm/issues/290