Struct fadroma::prelude::Uint128

pub struct Uint128(_);

A thin wrapper around u128 that is using strings for JSON encoding/decoding, such that the full u128 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.

Examples

Use from to create instances of this and u128 to get the value out:

let a = Uint128::from(123u128);
assert_eq!(a.u128(), 123);

let b = Uint128::from(42u64);
assert_eq!(b.u128(), 42);

let c = Uint128::from(70u32);
assert_eq!(c.u128(), 70);

Implementations

impl Uint128

pub const MAX: Uint128 = Self(u128::MAX)

pub const fn new(value: u128) -> Uint128

Creates a Uint128(value).

This method is less flexible than from but can be called in a const context.

pub const fn zero() -> Uint128

Creates a Uint128(0)

pub const fn one() -> Uint128

Creates a Uint128(1)

pub const fn u128(&self) -> u128

Returns a copy of the internal data

pub const fn to_be_bytes(self) -> [u8; 16]

Returns a copy of the number as big endian bytes.

pub const fn to_le_bytes(self) -> [u8; 16]

Returns a copy of the number as little endian bytes.

pub const fn is_zero(&self) -> bool

pub fn pow(self, exp: u32) -> Uint128

pub fn multiply_ratio<A, B>(&self, numerator: A, denominator: B) -> Uint128where
A: Into<u128>,
B: Into<u128>,

Returns self * numerator / denominator

pub fn checked_multiply_ratio<A, B>(
&self,
numerator: A,
denominator: B
) -> Result<Uint128, CheckedMultiplyRatioError>where
A: Into<u128>,
B: Into<u128>,

Returns self * numerator / denominator

pub fn full_mul(self, rhs: impl Into<u128>) -> Uint256

Multiplies two u128 values without overflow, producing an Uint256.

Examples

use cosmwasm_std::Uint128;

let a = Uint128::MAX;
let result = a.full_mul(2u32);
assert_eq!(result.to_string(), "680564733841876926926749214863536422910");

pub fn checked_add(self, other: Uint128) -> Result<Uint128, OverflowError>

pub fn checked_sub(self, other: Uint128) -> Result<Uint128, OverflowError>

pub fn checked_mul(self, other: Uint128) -> Result<Uint128, OverflowError>

pub fn checked_pow(self, exp: u32) -> Result<Uint128, OverflowError>

pub fn checked_div(self, other: Uint128) -> Result<Uint128, DivideByZeroError>

pub fn checked_div_euclid(
self,
other: Uint128
) -> Result<Uint128, DivideByZeroError>

pub fn checked_rem(self, other: Uint128) -> Result<Uint128, DivideByZeroError>

pub fn wrapping_add(self, other: Uint128) -> Uint128

pub fn wrapping_sub(self, other: Uint128) -> Uint128

pub fn wrapping_mul(self, other: Uint128) -> Uint128

pub fn wrapping_pow(self, other: u32) -> Uint128

pub fn saturating_add(self, other: Uint128) -> Uint128

pub fn saturating_sub(self, other: Uint128) -> Uint128

pub fn saturating_mul(self, other: Uint128) -> Uint128

pub fn saturating_pow(self, other: u32) -> Uint128