Struct fadroma::prelude::Uint256

pub struct Uint256(_);

An implementation of u256 that is using strings for JSON encoding/decoding, such that the full u256 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.

Examples

Use from to create instances out of primitive uint types or new to provide big endian bytes:

let a = Uint256::from(258u128);
let b = Uint256::new([
    0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8,
    0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8,
    0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8,
    0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 1u8, 2u8,
]);
assert_eq!(a, b);

Implementations

impl Uint256

pub const MAX: Uint256 = Uint256(U256::MAX)

pub const fn new(value: [u8; 32]) -> Uint256

Creates a Uint256(value) from a big endian representation. It’s just an alias for Uint256::from_be_bytes.

This method is less flexible than from but can be called in a const context.

pub const fn zero() -> Uint256

Creates a Uint256(0)

pub const fn one() -> Uint256

Creates a Uint256(1)

pub const fn from_be_bytes(data: [u8; 32]) -> Uint256

pub const fn from_le_bytes(data: [u8; 32]) -> Uint256

pub const fn from_u128(num: u128) -> Uint256

A conversion from u128 that, unlike the one provided by the From trait, can be used in a const context.

pub const fn from_uint128(num: Uint128) -> Uint256

A conversion from Uint128 that, unlike the one provided by the From trait, can be used in a const context.

pub const fn to_be_bytes(self) -> [u8; 32]

Returns a copy of the number as big endian bytes.

pub const fn to_le_bytes(self) -> [u8; 32]

Returns a copy of the number as little endian bytes.

pub const fn is_zero(&self) -> bool

pub fn pow(self, exp: u32) -> Uint256

pub fn multiply_ratio<A, B>(&self, numerator: A, denominator: B) -> Uint256where
A: Into<Uint256>,
B: Into<Uint256>,

Returns self * numerator / denominator

pub fn checked_multiply_ratio<A, B>(
&self,
numerator: A,
denominator: B
) -> Result<Uint256, CheckedMultiplyRatioError>where
A: Into<Uint256>,
B: Into<Uint256>,

Returns self * numerator / denominator

pub fn full_mul(self, rhs: impl Into<Uint256>) -> Uint512

Multiplies two u256 values without overflow, producing an Uint512.

Examples

use cosmwasm_std::Uint256;

let a = Uint256::MAX;
let result = a.full_mul(2u32);
assert_eq!(
    result.to_string(),
    "231584178474632390847141970017375815706539969331281128078915168015826259279870",
);

pub fn checked_add(self, other: Uint256) -> Result<Uint256, OverflowError>

pub fn checked_sub(self, other: Uint256) -> Result<Uint256, OverflowError>

pub fn checked_mul(self, other: Uint256) -> Result<Uint256, OverflowError>

pub fn checked_pow(self, exp: u32) -> Result<Uint256, OverflowError>

pub fn checked_div(self, other: Uint256) -> Result<Uint256, DivideByZeroError>

pub fn checked_rem(self, other: Uint256) -> Result<Uint256, DivideByZeroError>

pub fn checked_shr(self, other: u32) -> Result<Uint256, OverflowError>

pub fn checked_shl(self, other: u32) -> Result<Uint256, OverflowError>

pub fn saturating_add(self, other: Uint256) -> Uint256

pub fn saturating_sub(self, other: Uint256) -> Uint256

pub fn saturating_mul(self, other: Uint256) -> Uint256