Module fadroma::scrt

Secret Network specific utilities and contracts that are commonly used when developing smart contracts for it. Feature flag: scrt

Modules

permit

Authentication using SNIP-24 query permits. Feature flag: permit

snip20

vk

Viewing key authentication. Feature flag: vk

Constants

BLOCK_SIZE

Default Secret Network message padding size.

Functions

pad_response

Pad the given response using space_pad the default BLOCK_SIZE.

space_pad

Take a Vec and pad it up to a multiple of block_size using spaces at the end.

to_cosmos_msg

Creates a new WasmMsg::Execute using the provided msg and padding it to BLOCK_SIZE.