Module fadroma::scrt::snip20::contract

Customizable implementation of a SNIP-20 token. Feature flag: snip20-impl

If you simply need a vanilla SNIP-20 token you only need to look at the instantiate, execute and query functions. Call those from the respective entry points of your contract and pass DefaultImpl as the last parameter.

All other functions and modules are exposed for convenience so that you don’t need to copy code from here in order to make some changes to a method.

The the only methods which you might want to customize are Snip20::symbol_validation and Snip20::name_range which specify what characters can the token symbol be consisted of and between how many characters long can the token name be.

Modules

msg

receiver

state

transaction_history

Structs

DefaultImpl

The vanilla implementation of the SNIP-20 standard. Pass this to the instantiate, execute and query entry points if you don’t need to make any modifications.

SymbolValidation

Defines the length, letter casing and any special characters that are allowed as the symbol for the token.

Traits

Snip20

Functions

add_receiver_api_callback

assert_valid_name

assert_valid_symbol

execute

The execute entry point of the SNIP-20 contract.

instantiate

The instantiate entry point of the SNIP-20 contract.

mint_impl

perform_transfer

query

The query entry point of the SNIP-20 contract.

send_from_impl

send_impl

transfer_from_impl

transfer_impl

use_allowance