Struct fadroma::scrt::vk::ViewingKey

pub struct ViewingKey(pub String);

Represents a viewing key string which is provided as unverified input to a query.

PartialEq is intentionally not implemented on this type in order to prevent from using it to check viewing keys that way. You should convert to ViewingKeyHashed (using ViewingKey::to_hashed) and then call ViewingKeyHashed::check which performs a specialized constant time equality comparison. ViewingKey::check and ViewingKey::check_hashed are also provided for convenience.

Tuple Fields
0: String
Implementations

impl ViewingKey

pub fn new(env: &Env, info: &MessageInfo, seed: &[u8], entropy: &[u8]) -> Self

pub fn to_hashed(&self) -> ViewingKeyHashed

Converts the viewing key to SHA-256 hash representation.

pub fn as_bytes(&self) -> &[u8]

pub fn check(&self, other: &ViewingKey) -> bool

Converts both self and other to ViewingKeyHashed and compares them for equality using a constant time function.

pub fn check_hashed(&self, hashed: &ViewingKeyHashed) -> bool

Converts self to ViewingKeyHashed and compares it to other for equality using a constant time function.