Function fadroma::secret_toolkit::snip721::handle::remove_minters_msg

pub fn remove_minters_msg(
    minters: Vec<HumanAddr, Global>, 
    padding: Option<String>, 
    block_size: usize, 
    callback_code_hash: String, 
    contract_addr: HumanAddr
) -> Result<CosmosMsg<Empty>, StdError>

Returns a StdResult used to execute RemoveMinters

Arguments

minters - list of addresses that are no longer allowed to mint

padding - Optional String used as padding if you don’t want to use block padding

block_size - pad the message to blocks of this size

callback_code_hash - String holding the code hash of the contract being called

contract_addr - address of the contract being called