Crate fadroma::secret_toolkit::utils

Modules

calls

padding

Traits

HandleCallback

A trait marking types that define the handle message(s) of a contract

InitCallback

A trait marking types that define the instantiation message of a contract

Query

A trait marking types that define the query message(s) of a contract

Functions

pad_handle_result

Pad the data and logs in a HandleResult to the block size, with spaces.

pad_query_result

Pad a QueryResult with spaces

space_pad

Take a Vec and pad it up to a multiple of block_size, using spaces at the end.