Struct fadroma::storage::iterable::IterableStorage

pub struct IterableStorage<T: DeserializeOwned + Serialize, K: Key> { /* private fields */ }

Stores items in a way that allows for iterating over them in a sequential order just like a Vec. It’s also possible to retrieve or update inidividual items based on their index.

Implementations

impl<T: DeserializeOwned + Serialize, K: Key> IterableStorage<T, K>

pub fn new(ns: K) -> Self

Creates an instance for the given namespace. The following namespaces are reserved by IterableStorage:

ns + “index”

ns + N - where N is a number

pub fn iter<'storage>(
&self,
storage: &'storage dyn Storage
) -> StdResult<Iter<'storage, T>>

Returns an iterator that iterates through the stored elements in a sequential order.

Examples

use fadroma::storage::{iterable::IterableStorage, CompositeKey};
use fadroma::cosmwasm_std::testing::mock_dependencies;
 
let mut deps = mock_dependencies();
let s = deps.as_mut().storage;
 
let key = CompositeKey::new(&[b"numbers"]);
let mut storage = IterableStorage::<u8, _>::new(key);
 
storage.push(s, &1).unwrap();
storage.push(s, &2).unwrap();
 
let mut iter = storage.iter(s).unwrap();
assert_eq!(iter.next().unwrap(), Ok(1));
assert_eq!(iter.next().unwrap(), Ok(2));
assert_eq!(iter.next(), None)

pub fn push(&mut self, storage: &mut dyn Storage, value: &T) -> StdResult<u64>

Returns the index at which the item is stored at.

Examples

use fadroma::storage::{iterable::IterableStorage, CompositeKey};
use fadroma::cosmwasm_std::testing::mock_dependencies;
 
let mut deps = mock_dependencies();
let s = deps.as_mut().storage;
 
let key = CompositeKey::new(&[b"numbers"]);
let mut storage = IterableStorage::<u8, _>::new(key);
 
let index = storage.push(s, &1).unwrap();
assert_eq!(index, 0);
 
let index = storage.push(s, &2).unwrap();
assert_eq!(index, 1);

pub fn pop(&mut self, storage: &mut dyn Storage) -> StdResult<()>

Removes the item at the end of the collection. Does not return the removed element because that requires a storage read.

Examples

use fadroma::storage::{iterable::IterableStorage, CompositeKey};
use fadroma::cosmwasm_std::testing::mock_dependencies;
 
let mut deps = mock_dependencies();
let s = deps.as_mut().storage;
 
let key = CompositeKey::new(&[b"numbers"]);
let mut storage = IterableStorage::<u8, _>::new(key);
 
storage.push(s, &1).unwrap();
assert_eq!(storage.len(s).unwrap(), 1);
 
storage.pop(s).unwrap();
assert_eq!(storage.len(s).unwrap(), 0);

pub fn get_at(&self, storage: &dyn Storage, index: u64) -> StdResult<Option<T>>

Retruns the element stored at the given index or None if the index is out of bounds.

Examples

use fadroma::storage::{iterable::IterableStorage, CompositeKey};
use fadroma::cosmwasm_std::testing::mock_dependencies;
 
let mut deps = mock_dependencies();
let s = deps.as_mut().storage;
 
let key = CompositeKey::new(&[b"numbers"]);
let mut storage = IterableStorage::<u8, _>::new(key);
 
storage.push(s, &1).unwrap();
 
assert_eq!(storage.get_at(s, 0).unwrap(), Some(1));
assert_eq!(storage.get_at(s, 1).unwrap(), None);

pub fn update_at<F>(
&self,
storage: &mut dyn Storage,
index: u64,
update: F
) -> StdResult<Option<T>>where
F: FnOnce(T) -> StdResult<T>,

Returns the value returned by the provided update closure or None if nothing is stored at the given index.

Examples

use fadroma::storage::{iterable::IterableStorage, CompositeKey};
use fadroma::cosmwasm_std::testing::mock_dependencies;
 
let mut deps = mock_dependencies();
let s = deps.as_mut().storage;
 
let key = CompositeKey::new(&[b"numbers"]);
let mut storage = IterableStorage::<u8, _>::new(key);
 
storage.push(s, &1).unwrap();
 
let add_one = |mut x| {
    x += 1;
    Ok(x)
};
 
let updated_val = storage.update_at(s, 0, add_one).unwrap();
assert_eq!(updated_val, Some(2));
assert_eq!(storage.get_at(s, 0).unwrap(), Some(2));
 
let updated_val = storage.update_at(s, 1, add_one).unwrap();
assert_eq!(updated_val, None);

pub fn swap_remove(
&mut self,
storage: &mut dyn Storage,
index: u64
) -> StdResult<Option<T>>

Removes the element at the given index. The removed element is replaced by the last element in the storage. Does not preserve ordering. Returns the item that was swapped if such was necessary. Return an error if the index is out of bounds.

Examples

use fadroma::storage::{iterable::IterableStorage, CompositeKey};
use fadroma::cosmwasm_std::{
    StdError,
    testing::mock_dependencies
};
 
let mut deps = mock_dependencies();
let s = deps.as_mut().storage;
 
let key = CompositeKey::new(&[b"numbers"]);
let mut storage = IterableStorage::<u8, _>::new(key);
 
storage.push(s, &1).unwrap();
storage.push(s, &2).unwrap();
 
let removed = storage.swap_remove(s, 0).unwrap();
// We had to move the number 2 to index 0 so this is what is returned.
assert_eq!(removed, Some(2));
assert_eq!(storage.len(s).unwrap(), 1);
 
let removed = storage.swap_remove(s, 0).unwrap();
// We removed the last element so no reordering was necessary.
assert_eq!(removed, None);
assert_eq!(storage.len(s).unwrap(), 0);
 
let err = storage.swap_remove(s, 0).unwrap_err();
assert_eq!(err, StdError::generic_err("IterableStorage: index out of bounds."));

pub fn len(&self, storage: &dyn Storage) -> StdResult<u64>

Returns the number of element currently stored.

Examples

use fadroma::storage::{iterable::IterableStorage, CompositeKey};
use fadroma::cosmwasm_std::testing::mock_dependencies;
 
let mut deps = mock_dependencies();
let s = deps.as_mut().storage;
 
let key = CompositeKey::new(&[b"numbers"]);
let mut storage = IterableStorage::<u8, _>::new(key);
 
storage.push(s, &1).unwrap();
assert_eq!(storage.len(s).unwrap(), 1);
 
storage.pop(s).unwrap();
assert_eq!(storage.len(s).unwrap(), 0);