Trait fadroma::storage::Namespace

pub trait Namespace {
    const NAMESPACE: &'static [u8];
}

Represents a namespace, usually acting as a prefix to a dynamically generated key. We only do this so that we can have strongly typed keys and storage types. Use the namespace macro to generate one.

Examples

use fadroma::storage::Namespace;
 
fadroma::namespace!(MyNamespace, b"ns_bytes");
assert_eq!(MyNamespace::NAMESPACE, b"ns_bytes");

Required Associated Constants

const NAMESPACE: &'static [u8]