Module fadroma::killswitch

Emergency pause and termination of contracts. You MUST implement admin in your contract if you want to use this module.

Structs

ContractStatus

Current state of a contract w/ optional description and pointer to new version.

DefaultImpl

KillswitchNs

Enums

ContractStatusLevel

Possible states of a contract.

ExecuteMsg

QueryMsg

Constants

STORE

Traits

Killswitch

Requires the admin component in order to check for admin.

Functions

assert_can_set_status

Fail if trying to return from ContractStatusLevel::Migrating status.

assert_is_operational

Fail if the current contract status level is other than ContractStatusLevel::Operational.

execute

is_operational

Returns false if the current contract status level is other than ContractStatusLevel::Operational.

query

set_status

Store a new contract status. Requires the admin component in order to check for admin.