Module fadroma::killswitch

Feature flag: killswitch Emergency pause and termination of contracts.

Structs

ContractStatus

Current state of a contract w/ optional description and pointer to new version

DefaultImpl

Enums

ContractStatusLevel

Possible states of a contract.

HandleMsg

QueryMsg

Constants

PREFIX

Traits

Killswitch

Requires the admin component in order to check for admin.

Functions

can_set_status

Fail if trying to return from Migrating status.

get_status

Return the current contract status. Defaults to operational if nothing was stored.

handle

is_operational

Fail if the current contract status level is other than Operational.

load

query

save

set_status

Store a new contract status. Requires the admin component in order to check for admin.