Module fadroma::storage

Utilities for interacting with the native key-value storage.

Re-exports

pub use namespace::*;

Modules

namespace

Macros

stateful

Because Rust can’t yet genericize over mutability, this macro can be used to implement the same readonly methods twice - once for &S and for &mut S.

Structs

IterableStorage

Stores items in a way that allows for iterating over them.

StorageIterator

Traits

Readonly

Trait for actor that operates in a context with readonly access to the storage.

Writable

Trait for actor that operates in a context with mutable storage

Functions

concat

Concatenate a namespace and a key to get a namespaced key.

load

Load something from the storage.

ns_load

Load the value of a namespaced key.

ns_remove

Remove the value of a namespaced key from the storage.

ns_save

Save something to the storage under a namespaced key.

remove

Remove something from the storage.

save

Save something to the storage.